Add the ability to search for post offices
Created an axios post request to get data about post offices from the database.
Created new actions in officeSlicer.
When you click on the View Offices button, a request is sent to the server to search for available offices.
Added pagination to reduce the load on requests and improve the design.
setPageCount was created to calculate the number of pagination pages.