Create features for HistoryView component
A historyList property has been created in the trackingSlicer to store an array of previously viewed TTH Numbers. The setHistory method was created, with which you can change the array of viewed TTH numbers and save them in localStorage using the history key.
The showTThInfo method can also change the historyList and localStorage array.